Update interpreted message
the interpreted message in the error log was not correctly shown. Since The message variable, which is used as step_log, is a list of strings. When passed to interpret_gwas_workflow_error, it should be a single string, not a list. However, if message is accidentally passed as a list, in conditions in interpret_gwas_workflow_error will fail, causing show_error to remain "".